+++ src/lib/py/subcommands.py
@@ -1,5 +1,7 @@
 import Convert2MzML
+import DownloadData
 
 SUBCOMMANDS = [
     Convert2MzML.Convert2MzML,
+    DownloadData.DownloadData,
 ]