+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
+## 2026-01-06
+
+- Add command DoenloadData
+
 ## 2025-12-22
 
 - Add analysis to determine whether the data is non-overlapping.